backend/requirements.txt

fastapi
uvicorn[standard]
ultralytics
Pillow
numpy
opencv-python
python-multipart
